+++ smol-bytes/src/lib.rs
@@ -58,6 +58,10 @@ mod graphql;
 #[cfg_attr(docsrs, doc(cfg(feature = "pyo3")))]
 mod python;
 
+#[cfg(feature = "sqlx")]
+#[cfg_attr(docsrs, doc(cfg(feature = "sqlx")))]
+mod sql;
+
 #[cfg(feature = "wasm")]
 mod wasm;
 #[cfg(feature = "wasm")]